Have rel=opener prevent performance related browsing context group swaps
In "obtain a browsing context to use for a navigation response," a user
agent could swap browsing context groups, even when not needed for
security, when there are no other contexts that could script the window
that's navigating. This is however still a web visible behavior
difference, so we offer the use of rel=opener as a mechanism to prevent
these kinds of swaps.

This spec update is based on the following proposal:
https://github.com/explainers-by-googlers/future-browsing-context-group-dependency-hint